requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfminer.six==20231228
Pillow==10.0.1
numpy==1.24.3
opencv-python==4.6.0.66
paddlepaddle==2.5.1
paddleocr==2.7.0.3
python-magic==0.4.27
PyPDF2==3.0.1
PyMuPDF<1.21.0
requests==2.31.0
psutil==5.9.5
weasyprint==60.2
mistune==3.0.2
pypandoc==1.13
markdown==3.6
pydyf==0.10.0
